requirements.txt
pycryptodome
pytest
numpy
matplotlib
